split up CallCredentials.cs into multiple files
